Refactoring and fixed functionality of some menu options

+++ capstone/src/main/java/com/techelevator/Candy.java
@@ -1,7 +1,9 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Candy extends Product{
-    public Candy(String name, double price) {
+    public Candy(String name, BigDecimal price) {
         super(name, price);
     }
