+++ capstone/src/main/java/com/techelevator/Chip.java
@@ -1,7 +1,9 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Chip extends Product{
-    public Chip(String name, double price) {
+    public Chip(String name, BigDecimal price) {
         super(name, price);
     }
