+++ capstone/src/main/java/com/techelevator/Drink.java
@@ -1,7 +1,9 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public class Drink extends Product {
-    public Drink(String name, double price) {
+    public Drink(String name, BigDecimal price) {
         super(name, price);
     }
