+++ capstone/src/main/java/com/techelevator/Product.java
@@ -1,11 +1,13 @@
 package com.techelevator;
 
+import java.math.BigDecimal;
+
 public abstract class Product {
     private String name;
-    private double price;
+    private BigDecimal price;
     private int stock = 5;
 
-    public Product(String name, double price) {
+    public Product(String name, BigDecimal price) {
         this.name = name;
         this.price = price;
     }
@@ -14,7 +16,7 @@ public abstract class Product {
         return name;
     }
 
-    public double getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
